+++ pang/src/grammar/macros.rs
@@ -8,8 +8,8 @@
 /// use pang::{grammar, exp, nt, t};
 ///
 /// let grammar = grammar! {
-///     "start" => vec![exp(vec![nt("digit"), t("+"), nt("digit")])],
-///     "digit" => vec![exp(vec![t("0"), t("1"), t("2"), t("3"), t("4")])],
+///     "start" => [exp([nt("digit"), t("+"), nt("digit")])],
+///     "digit" => [exp([t("0"), t("1"), t("2"), t("3"), t("4")])],
 /// };
 /// ```
 #[macro_export]
@@ -23,7 +23,7 @@ macro_rules! grammar {
         {
             let mut grammar = $crate::grammar::Grammar::new();
             $(
-                grammar.insert($name.to_string(), $body);
+                grammar.insert($name.to_string(), $body.into());
             )*
             grammar
         }